refactor(communication): rename Discord message modules for clearer intent
