+++ service/communication/services/__init__.py
@@ -1,4 +1,4 @@
-from communication.services.discord import DiscordService
+from communication.services.discord_message import DiscordService
 from communication.services.email import EmailService
 
 __all__ = ["DiscordService", "EmailService"]
